fire focus events before injecting password
usability has made more and more sites do scripting
stuff on focusin/focusout of the password field to
give some sort of feedback on the password.

opening the popup will emit a focusout which will
then trigger this behaviour and show "invalid password"
or similar.

By issuing a focus event before injecting the password,
sites that hide that message on focusin will properly
do so.

issue #40