- browser_action: remove window-type filter on chrome.tabs.query lookup.
Chrome 45 added a new window type, 'app' - used when a website is set up as a pseudo-PWA.
This breaks masterpassword when using a pseudo-PWA.

Since there is no way to specify filtering on 'window' || 'app', the filter was removed completely.
I believe this is more succinct than firing one chrome.tabs.query for window, then trying app if the former fails.