Some formatting tidyups to help merging in future.
Areas targetted are ones likely to change when new features are added.
 - Arrays: Make sure one entry per line, and always trail with a comma
 - CSS selectors: One selector per line, a no-op fake selector to ensure every real selector ends with a trailing comma

Both of the above mean when doing merges, the only lines changed are ones added/removed - not the final one in the list just because a comma was added to extend the list.

index.html: Fixed up some indenting to make editing in future clearer.

(cherry picked from merge request GH-68 )