get second-level domain from publicsuffixlist
When suggesting a sitename we try to find the "significant"
part of the url. for www.google.com that would be google.com,
but just keeping the two last parts (or removing the first one)
fail too often. amazon.co.uk is one example.

Further, each TLD has it's own policy here, so an algorithmic
approch is bound to fail.  https://publicsuffix.org/ tries
to gather all possible SLD's. it might not be perfect, but
better than what we have (hardcoding a couple like (com|edu|
co).*

The list is rather large, but with some clever(?) tricks
we can get it down to an acceptable size:

Going a bit crazy here. Browsers don't support gzip/deflate data yet
(waiting for the Compression Streams API) and other compression
schemes where reasonable libs are available simply don't cut it
on the compression rate.

in the mean time, png is lossless and deflate compression -
exactly what we need  :)   So this patch pre-process theh PSL
list for easy lookup (and removes a lot of reduntant text) and
export the result as a json dictionary.

this is then converted to png by imagemagick.

The browser loads the image, we access the pixel values and end
up with our desired json dict.

Issue #69