Store (on request) mpw state instead of masterkey
If user has configured it, the key is stored in memory
so he/she does not have to type the master key every
time the popup opens.

When possible, this commit will change that to store the
mpw state (scrypt hash) instead of the masterkey, so
we pay the scrypt runtime penalty only once. Beneficial
on weak(er) devices like phones.

Also, trying to avoid running the scrypt (asm.js)
stuff in content-scripts in future inline popup, making
this change a requisite.