fix issue #79
Privacy conscious users might venture into privacy.resistFingerprinting
that effectively disables media queries for screen size

This as the rather annoying effect that the popup window is not displayed
since it needs a fixed size. On android however, the popup is a full page and
we want to use all available space.

Setting a min-height should be a reasonable compromise, hopefully not
pushing stuff "below the fold" on mobile.